Refatoração do mapa e mais funcionalidades
